Wrote 3-errors
