+++ text/appetizers/4-events.md
@@ -43,4 +43,32 @@ The syntax for `deposit_event` now takes the `RawEvent` type because it is gener
 Self::deposit_event(RawEvent::EmitInput(user, new_number));
 ```
 
+*NOTE*: The event described above only wraps `u32` values. If we want/need the `Event` type to contain multiple types from our runtime, then the `decl_event` would use the following syntax
+
+```rust, ignore
+decl_event!(
+    pub enum Event<T> {
+        ...
+    }
+)
+```
+
+In some cases, the `where` clause can be used to specify type aliasing for more readable code
+
+```rust, ignore
+decl_event!(
+    pub enum Event<T>
+    where
+        Balance = BalanceOf<T>,
+        <T as system::Trait>::AccountId,
+        <T as system::Trait>::BlockNumber,
+        <T as system::Trait>::Hash,
+    {
+        FakeEvent1(AccountId, Hash, BlockNumber),
+        FakeEvent2(AccountId, Balance, BlockNumber),
+    }
+)
+```
+
+
 *See the next example to use the simple event syntax in the context of verifying successful execution of an [adding machine](./adder.md)*
